+++ lib/docking.dart
@@ -1,4 +1,5 @@
 library docking;
 
 export 'src/layout/docking_layout.dart';
+export 'src/on_item_selection.dart';
 export 'src/docking.dart';