+++ lib/src/on_item_selection.dart
@@ -1,0 +1,4 @@
+import 'package:docking/src/layout/docking_layout.dart';
+
+/// Event that will be triggered when a [DockingItem] selection is changed.
+typedef OnItemSelection = Function(DockingItem item);